# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Central de Serviços API
  version: 1.0.0
  description: API REST da Central de Serviços v1. Datas são UTC em ISO-8601.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SERVICE_DESK_SESSION
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Requisição inválida.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    Unauthorized:
      description: Sessão ausente, inválida ou expirada.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    Forbidden:
      description: Perfil ou escopo insuficiente.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    Conflict:
      description: Conflito de unicidade ou versão otimista.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    Unprocessable:
      description: Regra de negócio não satisfeita.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
  schemas:
    Role:
      type: string
      enum: [REQUESTER, AGENT, MANAGER, ADMIN]
...
    AgendaItemKind:
      type: string
      enum: [INTERNAL_DEMAND, INSTITUTION_EVENT]
    AgendaItemStatus:
      type: string
      enum: [PENDING, COMPLETED]
...
    Me:
      type: object
      required: [id, email, displayName, roles, passwordChangeRequired]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        roles:
          type: array
          uniqueItems: true
          items: { $ref: "#/components/schemas/Role" }
        passwordChangeRequired: { type: boolean }
...
    AgendaItem:
      type: object
      required: [id, kind, title, description, location, assigneeId, assigneeName, status, startAt, endAt, allDay, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/AgendaItemKind" }
        title: { type: string }
        description: { type: [string, "null"] }
        location: { type: [string, "null"] }
        assigneeId: { type: [string, "null"], format: uuid }
        assigneeName: { type: [string, "null"] }
        status:
          oneOf:
            - { $ref: "#/components/schemas/AgendaItemStatus" }
            - { type: "null" }
        startAt: { type: string, format: date-time, description: Início inclusivo em UTC. }
        endAt: { type: string, format: date-time, description: Término exclusivo em UTC. }
        allDay: { type: boolean }
        version: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    InstitutionEventInput:
      type: object
      required: [kind, title, startAt, endAt, allDay]
      properties:
        kind: { type: string, const: INSTITUTION_EVENT }
        title: { type: string, minLength: 1, maxLength: 160 }
        description: { type: [string, "null"], maxLength: 4000 }
        location: { type: [string, "null"], maxLength: 200 }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        allDay: { type: boolean }
      additionalProperties: false
    InternalDemandInput:
      type: object
      required: [kind, title, startAt, endAt, allDay]
      properties:
        kind: { type: string, const: INTERNAL_DEMAND }
        title: { type: string, minLength: 1, maxLength: 160 }
        description: { type: [string, "null"], maxLength: 4000 }
        assigneeId: { type: [string, "null"], format: uuid }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        allDay: { type: boolean }
      additionalProperties: false
    AgendaItemCreate:
      oneOf:
        - { $ref: "#/components/schemas/InstitutionEventInput" }
        - { $ref: "#/components/schemas/InternalDemandInput" }
      discriminator:
        propertyName: kind
        mapping:
          INSTITUTION_EVENT: "#/components/schemas/InstitutionEventInput"
          INTERNAL_DEMAND: "#/components/schemas/InternalDemandInput"
    AgendaItemUpdate:
      type: object
      required: [title, startAt, endAt, allDay, version]
      properties:
        title: { type: string, minLength: 1, maxLength: 160 }
        description: { type: [string, "null"], maxLength: 4000 }
        location: { type: [string, "null"], maxLength: 200 }
        assigneeId: { type: [string, "null"], format: uuid }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        allDay: { type: boolean }
        version: { type: integer, format: int64 }
...
    ProblemDetail:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        correlationId: { type: [string, "null"] }
        errors: { type: object, additionalProperties: { type: string } }
paths:
...
  /auth/session:
    post:
      security: [{ csrfHeader: [] }]
      summary: Cria uma sessão server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Sessão criada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { description: Limite de tentativas excedido. }
    delete:
      security: [{ sessionCookie: [], csrfHeader: [] }]
      summary: Encerra a sessão atual.
      responses:
        "204": { description: Sessão encerrada. }
  /auth/me:
    get:
      summary: Obtém o usuário autenticado.
      responses:
        "200":
          description: Sessão atual.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/password/change:
    post:
      security: [{ sessionCookie: [], csrfHeader: [] }]
      summary: Troca a senha do usuário autenticado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 8, maxLength: 128 }
      responses:
        "204": { description: Senha alterada. }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /auth/password/forgot:
    post:
      security: [{ csrfHeader: [] }]
      summary: Solicita recuperação sem revelar se a conta existe.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "204": { description: Solicitação processada. }
  /auth/password/reset:
    post:
      security: [{ csrfHeader: [] }]
      summary: Redefine a senha com token de uso único.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8, maxLength: 128 }
      responses:
        "204": { description: Senha redefinida. }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /agenda/items:
    get:
      summary: Lista itens sobrepostos ao período, filtrados pelo perfil no banco.
      description: MANAGER recebe eventos e demandas internas; REQUESTER recebe somente eventos institucionais.
      parameters:
        - name: start
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: end
          in: query
          required: true
          description: Limite exclusivo do período consultado.
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Itens visíveis em ordem cronológica.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/AgendaItem" } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    post:
      security: [{ sessionCookie: [], csrfHeader: [] }]
      summary: Cria evento institucional ou demanda interna; exige MANAGER.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgendaItemCreate" }
      responses:
        "201":
          description: Item criado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgendaItem" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /agenda/items/{id}:
    patch:
      security: [{ sessionCookie: [], csrfHeader: [] }]
      summary: Atualiza os campos do item sem alterar seu tipo; exige MANAGER.
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgendaItemUpdate" }
      responses:
        "200":
          description: Item atualizado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgendaItem" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Item não encontrado. }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    delete:
      security: [{ sessionCookie: [], csrfHeader: [] }]
      summary: Exclui permanentemente um item; exige MANAGER.
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - name: version
          in: query
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Item excluído. }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Item não encontrado. }
        "409": { $ref: "#/components/responses/Conflict" }
  /agenda/items/{id}/status:
    patch:
      security: [{ sessionCookie: [], csrfHeader: [] }]
      summary: Conclui ou reabre uma demanda interna; exige MANAGER.
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, version]
              properties:
                status: { $ref: "#/components/schemas/AgendaItemStatus" }
                version: { type: integer, format: int64 }
      responses:
        "200":
          description: Andamento atualizado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgendaItem" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Item não encontrado. }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
...